Code consolidation/DRY programming. Moved repeated code from DecayProcessor/EpgpService to a new factory class
