+++ BuzzBot/BuzzBot/Epgp/Extensions/ServiceCollectionExtensions.cs
@@ -19,6 +19,7 @@ namespace BuzzBot.Epgp.Extensions
                 .AddSingleton<DecayProcessor>()
                 .AddSingleton<IAliasEventAlerter, AliasEventAlerter>()
                 .AddScoped<IUserService, UserService>()
+                .AddScoped<IEpgpTransactionFactory, EpgpTransactionFactory>()
                 .AddSingleton<IRaidRepository, RaidRepository>();
 
             services.AddTransient<AhnQirajTempleItemMapper>();
