+++ BuzzBot/BuzzBot/Epgp/IEpgpTransactionFactory.cs
@@ -1,0 +1,9 @@
+﻿using BuzzBotData.Data;
+
+namespace BuzzBot.Epgp
+{
+    public interface IEpgpTransactionFactory
+    {
+        EpgpTransaction GetTransaction(EpgpAlias alias, int value, string memo, TransactionType transactionType);
+    }
+}
